Bunch of passive tree setup changes
- Created better folder system for passive files
- Seperated most passives into their own cs files for organization
- Changed most passives to use the Value variable which is also properly reflected in the hjson files
