+++ Content/Passives/Magic/SmallPassives/PercentIncreasedMaximumManaPassive.cs
@@ -1,0 +1,17 @@
+using PathOfTerraria.Common.Systems.ModPlayers;
+using PathOfTerraria.Common.Systems.PassiveTreeSystem;
+using PathOfTerraria.Common.Systems.TreeSystem;
+using PathOfTerraria.Core.Items;
+using Terraria.ModLoader.IO;
+
+namespace PathOfTerraria.Content.Passives;
+
+
+// {0}% Increased Maximum Mana
+internal class PercentIncreasedMaximumManaPassive : Passive
+{
+	public override void BuffPlayer(Player player)
+	{
+		player.statManaMax2 = (int)(player.statManaMax2 * (1.0f + (Value / 100.0f) * Level));
+	}
+}
