--- Content/Passives/MagicRangedPassives.cs
@@ -1,7 +1,0 @@
-﻿using PathOfTerraria.Common.Systems.PassiveTreeSystem;
-
-namespace PathOfTerraria.Content.Passives;
-
-internal class LongRangePassive : Passive
-{
-}
