+++ Content/Passives/Misc/AnchorPassive.cs
@@ -1,9 +1,7 @@
-﻿using PathOfTerraria.Common.Systems.ModPlayers;
 using PathOfTerraria.Common.Systems.PassiveTreeSystem;
 
 namespace PathOfTerraria.Content.Passives;
 
-internal class TestMasteryPassive: MasteryPassive
+internal class AnchorPassive : Passive
 {
-
 }
