+++ Content/Passives/Misc/JewelSocket.cs
