+++ Content/Passives/Misc/LifePassives.cs
@@ -7,7 +7,7 @@ internal class AddedLifePassive : Passive
 {
 	public override void BuffPlayer(Player player)
 	{
-		player.statLifeMax2 += 20 * Level;
+		player.statLifeMax2 += Value * Level;
 	}
 }
 
@@ -15,7 +15,7 @@ internal class LifeRegenPassive : Passive
 {
 	public override void BuffPlayer(Player player)
 	{
-		player.lifeRegen += 2 * Level;
+		player.lifeRegen += Value * Level;
 	}
 }
 
@@ -23,6 +23,6 @@ internal class IncreasedPotionRecoveryPassive : Passive
 {
 	public override void BuffPlayer(Player player)
 	{
-		player.GetModPlayer<UniversalBuffingPlayer>().UniversalModifier.PotionHealPower += Level * 0.5f;
+		player.GetModPlayer<UniversalBuffingPlayer>().UniversalModifier.PotionHealPower += Level * (Value/100.0f);
 	}
 }
