+++ Content/Passives/Misc/StartPassive.cs
@@ -1,0 +1,11 @@
+﻿using PathOfTerraria.Common.Systems.PassiveTreeSystem;
+
+namespace PathOfTerraria.Content.Passives;
+
+internal class StartPassive : Passive
+{
+	public override bool CanDeallocate(Player player)
+	{
+		return false;
+	}
+}
