+++ Content/Passives/Ranged/SmallPassives/ChanceToConsumeAmmoPassive.cs
@@ -1,0 +1,4 @@
+﻿using PathOfTerraria.Common.Systems.PassiveTreeSystem;
+
+namespace PathOfTerraria.Content.Passives;
+
