+++ Content/Passives/Ranged/SmallPassives/IncreasedRangedDamagePassive.cs
@@ -1,0 +1,16 @@
+using PathOfTerraria.Common.Systems.ModPlayers;
+using PathOfTerraria.Common.Systems.PassiveTreeSystem;
+using PathOfTerraria.Common.Systems.TreeSystem;
+using PathOfTerraria.Core.Items;
+using Terraria.ModLoader.IO;
+
+namespace PathOfTerraria.Content.Passives;
+
+
+internal class IncreasedRangedDamagePassive : Passive
+{
+	public override void BuffPlayer(Player player)
+	{
+		player.GetDamage(DamageClass.Ranged) += (Value/100.0f) * Level;
+	}
+}
