+++ Content/Passives/Summon/SmallPassives/IncreasedWhipSpeedPassive.cs
@@ -1,0 +1,13 @@
+using PathOfTerraria.Common.Systems.ModPlayers;
+using PathOfTerraria.Common.Systems.PassiveTreeSystem;
+using PathOfTerraria.Common.Systems.TreeSystem;
+using PathOfTerraria.Core.Items;
+using Terraria.ModLoader.IO;
+
+namespace PathOfTerraria.Content.Passives;
+
+
+internal class IncreasedWhipSpeedPassive : Passive
+{
+	
+}
