+++ Localization/ru-RU/Mods.PathOfTerraria.Passives.hjson
@@ -19,11 +19,6 @@ IncreasedMagicDamagePassive: {
 	Tooltip: Увеличивает ваш магический урон на 5% за уровень
 }
 
-IncreasedSummoningDamagePassive: {
-	Name: Мастерство призыва
-	Tooltip: Увеличивает урон ваших призывов на 5% за уровень
-}
-
 AddedLifePassive: {
 	Name: Укрепленная плоть
 	Tooltip: Увеличивает ваш максимальный запас здоровья на 20 за уровень
@@ -57,11 +52,6 @@ DamageReductionPassive: {
 DamageReflectionPassive: {
 	Name: Колючая внешность
 	Tooltip: Отражает дополнительно 10 урона за уровень врагам, наносящим контактный урон.
-}
-
-ChanceToConsumeAmmoPassive: {
-	Name: Секретный отсек
-	Tooltip: Шанс не расходовать боеприпасы: +5% за уровень
 }
 
 StartPassive: {
@@ -191,25 +181,35 @@ IncreasedJumpHeight: {
 
 PercentIncreasedMaximumManaPassive: {
 	// Name: Deep in Thought
-	// Tooltip: +5% Increased Maximum Mana
+	// Tooltip: +{0}% Increased Maximum Mana
 }
 
 IncreasedFireDamagePassive: {
 	// Name: Burning Fury
-	// Tooltip: +10% Increased Fire Damage
+	// Tooltip: +{0}% Increased Fire Damage
 }
 
 IncreasedLightningDamagePassive: {
 	// Name: Voltaic Surge
-	// Tooltip: +10% Increased Lightning Damage
+	// Tooltip: +{0}% Increased Lightning Damage
 }
 
 IncreasedColdDamagePassive: {
 	// Name: Chilling Precision
-	// Tooltip: +10% Increased Cold Damage
+	// Tooltip: +{0}% Increased Cold Damage
 }
 
 TestMasteryPassive: {
 	// Name: TestMasteryPassive
+	// Tooltip: ""
+}
+
+IncreasedWhipDamagePassive: {
+	// Name: IncreasedWhipDamagePassive
+	// Tooltip: ""
+}
+
+IncreasedWhipSpeedPassive: {
+	// Name: IncreasedWhipSpeedPassive
 	// Tooltip: ""
 }
